OneDrive/Desktop/Projects/AI powered Offline Exam Invigilator/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.26.0
scipy>=1.12.0
pyaudio>=0.2.13
SpeechRecognition>=3.10.0
python-multipart>=0.0.6
pygetwindow>=0.0.9
psutil>=5.9.6
python-dotenv>=1.0.0
pydantic>=2.5.0
Pillow>=10.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
